implement anim/realtime rig swapping at anim publish time
This is done so the animators can use a light and fast representation of the rig, which gets replaced by another representation including a realtime skeleton, matching the exported base .fbx.

The way this works is by splitting the rig into 3 components at rig publish time a.k.a. `extract_rig`:
- base .fbx file, containing the realtime skeleton and the geometries skinned to it
- animation .ma file, containing the animation rig, meant to be used only in Maya
- realtime .ma file, containing the animation rig AND the realtime skeleton, driven by the animation rig, so it can be cached out

And then at animation publish time a.k.a. `extract_pointcache`, before performing the extraction the animation rig .ma gets replaced by the realtime .ma.

NOTE!! Almost certainly due to my unfamiliarity with OpenPype I wasn't able to get the realtime .ma to export as an actual .ma and for the time being it is exporting to an .mb.