add layout subsets to sequence and level names in UE to support multiple layouts Since animations are now associated with a layout then we can have multiple layouts and import the animation only in the specified one.
Supporting that has meant that the shot levels and sequences now have the
_layout{subset} prefix e.g. {SHOT}_layoutMain_map, to differentiate between
different layouts.

The cameras are NOT associated with layouts, so they are imported in all
existing layouts, which means that we have multiple copies of the same
camera in the shot levels, but since the name of the camera is the same
we only have one instance in the sequencer.