Implement new method of handling sequences (as in parents of shots) The new method uses a single sequence Level and LevelSequence to represent anything that requires continuity across the whole sequence.
That level gets streamed into all shots upon loading the layout for that shot AND
the LevelSequence gets added as a subsequence.

Therefore, anything added to the sequence Level and LevelSequence will end up in all
of its shots with the full timeline of the sequence, so continuity dependend assets
such as simulations and crowds can be added to the one sequence Level and LevelSequence
and be present in all shots without continuity issues.